+++ packages/site/src/features/update-snap/Update.tsx
@@ -1,19 +1,14 @@
 import { FunctionComponent } from 'react';
 import { Button, ButtonGroup } from 'react-bootstrap';
 import { Result, Snap } from '../../components';
-import {
-  useGetSnapsQuery,
-  useInstallSnapMutation,
-  useInvokeMutation,
-} from '../../api';
+import { useGetSnapsQuery, useInstallSnapMutation } from '../../api';
 
 const UPDATE_SNAP_ID = 'npm:@metamask/test-snap-bip32';
-const UPDATE_SNAP_OLD_VERSION = '4.0.1';
-const UPDATE_SNAP_NEW_VERSION = '4.0.2';
+const UPDATE_SNAP_OLD_VERSION = '5.1.1';
+const UPDATE_SNAP_NEW_VERSION = '5.1.2';
 
 export const Update: FunctionComponent = () => {
   const [installSnap, { isLoading }] = useInstallSnapMutation();
-  const [invokeSnap, { isLoading: isInvokeLoading }] = useInvokeMutation();
   const { data: snaps } = useGetSnapsQuery();
 
   const currentVersion = snaps?.[UPDATE_SNAP_ID]?.version;
@@ -22,14 +17,6 @@ export const Update: FunctionComponent = () => {
     installSnap({
       snapId: UPDATE_SNAP_ID,
       version: UPDATE_SNAP_NEW_VERSION,
-    });
-  };
-
-  const handleTest = () => {
-    invokeSnap({
-      snapId: UPDATE_SNAP_ID,
-      method: 'confirm',
-      params: ['Hello, world!'],
     });
   };
 
@@ -48,19 +35,11 @@ export const Update: FunctionComponent = () => {
 
       <ButtonGroup>
         <Button
-          variant="primary"
           disabled={isLoading}
           onClick={handleUpdate}
           id="connectUpdateNew"
         >
           Update Snap
-        </Button>
-        <Button
-          variant="secondary"
-          onClick={handleTest}
-          disabled={isInvokeLoading}
-        >
-          Send Test
         </Button>
       </ButtonGroup>
     </Snap>