+++ packages/notification/CHANGELOG.md
@@ -6,6 +6,58 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.5.0]
+### Changed
+- No changes this release.
+
+## [5.4.0]
+### Changed
+- No changes this release.
+
+## [5.3.0]
+### Changed
+- No changes this release.
+
+## [5.2.0]
+### Changed
+- No changes this release.
+
+## [5.1.2]
+### Changed
+- No changes this release.
+
+## [5.1.1]
+### Changed
+- No changes this release.
+
+## [5.1.0]
+### Changed
+- No changes this release.
+
+## [5.0.3]
+### Changed
+- No changes this release.
+
+## [5.0.2]
+### Changed
+- No changes this release.
+
+## [5.0.1]
+### Changed
+- No changes this release.
+
+## [5.0.0]
+### Changed
+- Update snaps packages ([#141](https://github.com/MetaMask/test-snaps/pull/141))
+
+## [4.6.4]
+### Changed
+- No changes this release.
+
+## [4.6.3]
+### Changed
+- Update snaps-cli, snap shasums ([#136](https://github.com/MetaMask/test-snaps/pull/136))
+
 ## [4.6.2]
 ### Changed
 - No changes this release.
@@ -95,7 +147,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release ([#41](https://github.com/MetaMask/test-snaps/pull/41))
 
-[Unreleased]: https://github.com/MetaMask/test-snaps/compare/v4.6.2...HEAD
+[Unreleased]: https://github.com/MetaMask/test-snaps/compare/v5.5.0...HEAD
+[5.5.0]: https://github.com/MetaMask/test-snaps/compare/v5.4.0...v5.5.0
+[5.4.0]: https://github.com/MetaMask/test-snaps/compare/v5.3.0...v5.4.0
+[5.3.0]: https://github.com/MetaMask/test-snaps/compare/v5.2.0...v5.3.0
+[5.2.0]: https://github.com/MetaMask/test-snaps/compare/v5.1.2...v5.2.0
+[5.1.2]: https://github.com/MetaMask/test-snaps/compare/v5.1.1...v5.1.2
+[5.1.1]: https://github.com/MetaMask/test-snaps/compare/v5.1.0...v5.1.1
+[5.1.0]: https://github.com/MetaMask/test-snaps/compare/v5.0.3...v5.1.0
+[5.0.3]: https://github.com/MetaMask/test-snaps/compare/v5.0.2...v5.0.3
+[5.0.2]: https://github.com/MetaMask/test-snaps/compare/v5.0.1...v5.0.2
+[5.0.1]: https://github.com/MetaMask/test-snaps/compare/v5.0.0...v5.0.1
+[5.0.0]: https://github.com/MetaMask/test-snaps/compare/v4.6.4...v5.0.0
+[4.6.4]: https://github.com/MetaMask/test-snaps/compare/v4.6.3...v4.6.4
+[4.6.3]: https://github.com/MetaMask/test-snaps/compare/v4.6.2...v4.6.3
 [4.6.2]: https://github.com/MetaMask/test-snaps/compare/v4.6.1...v4.6.2
 [4.6.1]: https://github.com/MetaMask/test-snaps/compare/v4.6.0...v4.6.1
 [4.6.0]: https://github.com/MetaMask/test-snaps/compare/v4.5.0...v4.6.0